refactor(evm): centralize zone transaction validation
