+++ crates/evm/src/zone_evm/contract_creation.rs
@@ -1,10 +1,9 @@
-//! Zone transaction validation and contract creation runtime enforcement.
+//! Contract creation runtime enforcement.
 
 use alloy_evm::Database;
 use alloy_primitives::Address;
 use revm::{
     bytecode::opcode::{CREATE, CREATE2},
-    context::Transaction,
     interpreter::{
         Instruction, InstructionContext, InstructionResult,
         instructions::contract::create as revm_create, interpreter::EthInterpreter,
@@ -12,7 +11,7 @@ use revm::{
     },
 };
 use tempo_evm::evm::TempoEvm;
-use tempo_revm::{TempoInvalidTransaction, TempoTxEnv, evm::TempoContext};
+use tempo_revm::evm::TempoContext;
 use zone_primitives::constants::CONTRACT_DEPLOYER_ALLOWLIST;
 
 type ZoneInstructionCtx<'a, DB> = InstructionContext<'a, TempoContext<DB>, EthInterpreter>;
@@ -48,43 +47,10 @@ fn create<const IS_CREATE2: bool, DB: Database>(
     revm_create::<IS_CREATE2, EthInterpreter, TempoContext<DB>>(context)
 }
 
-/// Validates Zone transaction policies shared by pool admission and block execution.
-pub fn validate_transaction(
-    tx: &TempoTxEnv,
-    allowlist: &[Address],
-) -> Result<(), TempoInvalidTransaction> {
-    let has_eip7702_authorizations = !tx.inner.authorization_list.is_empty();
-    let has_tempo_authorizations = tx
-        .tempo_tx_env
-        .as_ref()
-        .is_some_and(|env| !env.tempo_authorization_list.is_empty());
-    if has_eip7702_authorizations || has_tempo_authorizations {
-        return Err(TempoInvalidTransaction::CallsValidation(
-            "authorization lists are not supported",
-        ));
-    }
-
-    if contract_creation_deployer(tx).is_some_and(|deployer| !allowlist.contains(&deployer)) {
-        return Err(TempoInvalidTransaction::CallsValidation(
-            "contract creation is not supported",
-        ));
-    }
-
-    Ok(())
-}
-
-fn contract_creation_deployer(tx: &TempoTxEnv) -> Option<Address> {
-    let creates = match tx.tempo_tx_env.as_ref() {
-        Some(aa) => aa.aa_calls.iter().any(|call| call.to.is_create()),
-        None => tx.kind().is_create(),
-    };
-    creates.then_some(tx.caller)
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::{L1OverlayDB, ZoneEvm};
+    use crate::{L1OverlayDB, ZoneEvm, zone_evm::validation::validate_transaction};
     use alloy_evm::{Evm, EvmEnv};
     use alloy_primitives::{Address, Bytes, TxKind, U256, bytes};
     use revm::{
@@ -99,7 +65,7 @@ mod tests {
     };
     use tempo_evm::{TempoBlockEnv, TempoHaltReason, TempoPoolValidationEvm};
     use tempo_primitives::transaction::Call;
-    use tempo_revm::{TempoBatchCallEnv, TempoTxEnv};
+    use tempo_revm::{TempoBatchCallEnv, TempoInvalidTransaction, TempoTxEnv};
     use zone_precompiles::test_utils::MockL1Reader as TestL1;
 
     type TestDb = CacheDB<EmptyDB>;
