+++ crates/evm/src/zone_evm/validation.rs
@@ -1,0 +1,40 @@
+//! Zone transaction policies shared by pool admission and block execution.
+
+use alloy_primitives::Address;
+use revm::context::Transaction;
+use tempo_revm::{TempoInvalidTransaction, TempoTxEnv};
+
+/// Validates transaction policies enforced by the Zone EVM.
+pub fn validate_transaction(
+    tx: &TempoTxEnv,
+    contract_deployer_allowlist: &[Address],
+) -> Result<(), TempoInvalidTransaction> {
+    let has_eip7702_authorizations = !tx.inner.authorization_list.is_empty();
+    let has_tempo_authorizations = tx
+        .tempo_tx_env
+        .as_ref()
+        .is_some_and(|env| !env.tempo_authorization_list.is_empty());
+    if has_eip7702_authorizations || has_tempo_authorizations {
+        return Err(TempoInvalidTransaction::CallsValidation(
+            "authorization lists are not supported",
+        ));
+    }
+
+    if contract_creation_deployer(tx)
+        .is_some_and(|deployer| !contract_deployer_allowlist.contains(&deployer))
+    {
+        return Err(TempoInvalidTransaction::CallsValidation(
+            "contract creation is not supported",
+        ));
+    }
+
+    Ok(())
+}
+
+fn contract_creation_deployer(tx: &TempoTxEnv) -> Option<Address> {
+    let creates = match tx.tempo_tx_env.as_ref() {
+        Some(aa) => aa.aa_calls.iter().any(|call| call.to.is_create()),
+        None => tx.kind().is_create(),
+    };
+    creates.then_some(tx.caller)
+}
